chore(build.gradle.kts): update Kotlin DSL plugin versions and remove unused plugins chore(Dependencies.kt): update Kotlin DSL version and Gradle Publish plugin version chore(dokka-storybook-plugin/build.gradle.kts): update Kotlin DSL plugin versions chore(fixers-plugin/build.gradle.kts): remove unused plugins chore(gradle.properties): update Gradle version to 8.2.1 chore(pom.gradle): add configuration for POM metadata chore(publishing_plugin.gradle): configure publishing of Maven artifacts to Sonatype chore(gradle-wrapper.properties): update Gradle version to 8.2.1 chore(sample-fragment/build.gradle.kts): update Kotlin DSL plugin version and add dependencies chore(sample/build.gradle.kts): update Kotlin DSL plugin version and add dependencies
The changes were made to update the Kotlin DSL plugin versions and remove unused plugins. The Gradle version was also updated to 8.2.1. Additionally, the POM metadata configuration and publishing of Maven artifacts to Sonatype were added. Dependencies were updated in the sample-fragment and sample projects. These changes were made to keep the project up to date with the latest versions and improve the build configuration.